Exercise content morph controls and share the semantic probe option binding
